feature/ECO-2504: CR Fix
