--- src/SprykerEco/Yves/CrefoPay/Theme/default/components/molecules/crefopay-form-loader/index.ts
@@ -1,3 +1,0 @@
-import './crefopay-form-loader.scss';
-import register from 'ShopUi/app/registry';
-export default register('crefopay-form-loader', () => import(/* webpackMode: "lazy" */'./crefopay-form-loader'));
