+++ src/SprykerEco/Yves/CrefoPay/Theme/default/components/molecules/crefopay-payment-form/crefopay-payment-form.twig
@@ -1,35 +1,30 @@
 {% extends model('component') %}
 
 {% define config = {
-    name: 'crefopay-form-loader',
-    tag: 'crefopay-form-loader'
+    name: 'crefopay-payment-form',
+    tag: 'crefopay-payment-form'
 } %}
 
 {% define data = {
     form: required,
     shopPublicKey: required,
     orderId: required,
-    registerPaymentActionUrl: required,
     secureFieldsLibraryUrl: required,
-    secureFieldsPlaceholders: {},
-    fields: []
-} %}
-
-{% if data.secureFieldsPlaceholders is not empty %}
-    {% set crefoPayConfig = {
-        url: data.registerPaymentActionUrl,
+    fields: [],
+    crefoPayConfig: {
+        url: required,
         placeholders: {
-            accountHolder: data.secureFieldsPlaceholders.accountHolder,
-            number: data.secureFieldsPlaceholders.number,
-            cvv: data.secureFieldsPlaceholders.cvv
+            accountHolder: required,
+            number: required,
+            cvv: required
         }
-    } %}
-{% endif %}
+    }
+} %}
 
 {% define attributes = {
     'shop-public-key': data.shopPublicKey,
     'order-id': data.orderId,
-    'crefo-pay-config': data.secureFieldsPlaceholders ? crefoPayConfig | json_encode() : '',
+    'crefo-pay-config': data.crefoPayConfig | json_encode(),
     'payment-form-selector': '#payment-form',
     'payment-instrument-id-selector': '[name="' ~ data.form.crefoPayCreditCard.children['paymentInstrumentId'].vars.full_name ~ '"]',
     'class-to-toggle': 'is-hidden',
@@ -60,7 +55,7 @@
         {% endfor %}
 
         <div class="{{ config.name }}__error {{ config.jsName }}__error spacing-y is-hidden">
-            {{ 'All fields are required.' | trans }}
+            {{ 'crefopay.required_notification' | trans }}
         </div>
     {% endblock %}
 {% endblock %}
