+++ src/SprykerEco/Yves/CrefoPay/Theme/default/components/molecules/crefopay-payment-form/index.ts
@@ -1,0 +1,3 @@
+import './crefopay-payment-form.scss';
+import register from 'ShopUi/app/registry';
+export default register('crefopay-payment-form', () => import(/* webpackMode: "lazy" */'./crefopay-payment-form'));
