+++ src/SprykerEco/Yves/CrefoPay/Theme/default/views/credit-card-3d/credit-card-3d.twig
@@ -15,9 +15,17 @@
     {% include molecule('crefopay-form-loader', 'CrefoPay') with {
         data: {
             form: data.form,
-            shopPublicKey: data.form.crefoPayCreditCard3D.vars.shopPublicKey,
-            orderId: data.form.crefoPayCreditCard3D.vars.orderID,
-            fields: crefopayFields
+            shopPublicKey: '',
+            orderId: '',
+            secureFieldsLibraryUrl: '',
+            crefoPayConfig: {
+                url: '',
+                placeholders: {
+                    accountHolder: '',
+                    number: '',
+                    cvv: ''
+                }
+            }
         }
     } only %}
 {% endblock %}
